fix start farm list
