+++ MainCore/Helper/Implementations/Base/RallypointHelper.cs
@@ -10,7 +10,7 @@ using System.Linq;
 
 namespace MainCore.Helper.Implementations.Base
 {
-    public abstract class RallypointHelper : IRallypointHelper
+    public class RallypointHelper : IRallypointHelper
     {
         protected readonly IDbContextFactory<AppDbContext> _contextFactory;
         protected readonly IChromeManager _chromeManager;
@@ -27,7 +27,21 @@ namespace MainCore.Helper.Implementations.Base
             _farmListParser = farmListParser;
         }
 
-        public abstract Result ClickStartFarm(int accountId, int farmId);
+        public Result ClickStartFarm(int accountId, int farmId)
+        {
+            var chromeBrowser = _chromeManager.Get(accountId);
+            var html = chromeBrowser.GetHtml();
+
+            var startButton = _farmListParser.GetStartButton(html, farmId);
+            if (startButton is null)
+            {
+                return Result.Fail(new Retry("Cannot found start button"));
+            }
+
+            var result = _generalHelper.Click(accountId, By.XPath(startButton.XPath));
+            if (result.IsFailed) return result.WithError(new Trace(Trace.TraceMessage()));
+            return Result.Ok();
+        }
 
         public Result EnterFarmListPage(int accountId, int villageId)
         {
