--- MainCore/Helper/Implementations/TravianOfficial/RallypointHelper.cs
@@ -1,40 +1,0 @@
-﻿using FluentResults;
-using MainCore.Errors;
-using MainCore.Helper.Interface;
-using MainCore.Parsers.Interface;
-using MainCore.Services.Interface;
-using Microsoft.EntityFrameworkCore;
-using OpenQA.Selenium;
-using System.Linq;
-
-namespace MainCore.Helper.Implementations.TravianOfficial
-{
-    public class RallypointHelper : Base.RallypointHelper
-    {
-        public RallypointHelper(IChromeManager chromeManager, IGeneralHelper generalHelper, IDbContextFactory<AppDbContext> contextFactory, IUpdateHelper updateHelper, IFarmListParser farmListParser) : base(chromeManager, generalHelper, contextFactory, updateHelper, farmListParser)
-        {
-        }
-
-        public override Result ClickStartFarm(int accountId, int farmId)
-        {
-            var chromeBrowser = _chromeManager.Get(accountId);
-            var html = chromeBrowser.GetHtml();
-
-            var farmNode = html.GetElementbyId($"raidList{farmId}");
-            if (farmNode is null)
-            {
-                return Result.Fail(new Retry("Cannot found farm node"));
-            }
-            var startNode = farmNode.Descendants("button")
-                                    .FirstOrDefault(x => x.HasClass("startButton"));
-            if (startNode is null)
-            {
-                return Result.Fail(new Retry("Cannot found start button"));
-            }
-
-            var result = _generalHelper.Click(accountId, By.XPath(startNode.XPath));
-            if (result.IsFailed) return result.WithError(new Trace(Trace.TraceMessage()));
-            return Result.Ok();
-        }
-    }
-}
